test: generalize the renderer
