+++ test/components/component-render.ts
@@ -1,0 +1,21 @@
+import { mountSuspended } from 'nuxt-vitest/utils'
+import path from 'path'
+
+export default async function (nameOrHtml: string, options: any, component: any) {
+  let html
+  const name = path.parse(component.__file).name
+  if (options === undefined) {
+    const app = {
+      template: nameOrHtml,
+      components: { [`U${name}`]: component }
+    }
+    const result = await mountSuspended(app)
+    html = result.html()
+  } else {
+    options.slots = options.slots || { default: () => 'label' }
+    options.slots.default = options.slots.default || (() => 'label')
+    const cResult = await mountSuspended(component, options)
+    html = cResult.html()
+  }
+  return html
+}
